fix(workers): bind interrupted-acquisition attempt cap as INT4, not INT8
The NONRANKED_ACQUISITION_INTERRUPT_MAX_ATTEMPTS error was serializing
as i64 against a $1::int placeholder, so the startup auto-ingester:discovery
pass failed with "error serializing parameter 0" and the nonranked
acquisition never ran. Parse as i32 so tokio-postgres binds INT4.

added:
- i32 parsing for the attempt fuse

Changelog: hidden

+++ src/backend-rust/src/workers/pipeline.rs
@@ -855,7 +855,7 @@ impl CanonicalIngestPipeline {
         // parks genuinely-unavailable matches instead of churning forever.
         let attempt_cap = std::env::var("NONRANKED_ACQUISITION_INTERRUPT_MAX_ATTEMPTS")
             .ok()
-            .and_then(|raw| raw.parse::<i64>().ok())
+            .and_then(|raw| raw.parse::<i32>().ok())
             .unwrap_or(6)
             .max(1);
         self.database.query_json(
